db-agent/app/requirement.txt
tenacity
pydantic>=2
requests
langchain
langgraph
fastapi
uvicorn
tenacity
python-dotenv
sqlalchemy
pymysql
psycopg2-binary